pyproject.toml
[build-system]
build-backend = "pdm.backend"
requires = [ "pdm-backend" ]

[project]
name = "autodarktheme"
version = "0.1.0"
license = { file = "LICENSE" }
requires-python = ">=3.12,<3.14"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyside6-essentials>=6.7.2",
]

[dependency-groups]
dev = [
    "coverage>=7.5.4",
    "mypy>=1.10.1",
    "pyproject-fmt>=2.11",
    "ruff>=0.5.1",
]

[tool.pdm]
distribution = true

[tool.pdm.build]
excludes = [ "tests" ]

[tool.pdm.scripts]
_.env = { PYTHONPATH = "src:tests" }
lint = { composite = [
    "pyproject-fmt pyproject.toml",
    "ruff format {args:.}",
    "ruff check {args:.}",
    "mypy {args:.}",
], keep_going = true }
tests = { composite = [
    "coverage run -m unittest discover --verbose --locals --failfast --catch --start-directory tests {args}",
    "coverage report",
    "coverage erase",
] }
start = { call = "autodarktheme.main:main" }

[tool.ruff]
line-length = 80
src = [ "src", "tests" ]
fix = true
format.quote-style = "single"
format.skip-magic-trailing-comma = true
format.docstring-code-format = true
lint.select = [ "ALL" ]
lint.ignore = [
    "ANN1",
    "COM812",
    "D1",
    "D203",
    "D213",
    "ISC001",
    "PT009",
    "PT027",
    "Q001",
    "Q002",
    "TD002",
    "TD003",
]
lint.extend-per-file-ignores."stubs/**/*.pyi" = [
    "FBT001",
    "FBT002",
    "N802",
    "N803",
    "N815",
    "PLR0913",
]
lint.flake8-annotations.allow-star-arg-any = true
lint.flake8-quotes.docstring-quotes = "single"
lint.flake8-quotes.inline-quotes = "single"
lint.flake8-quotes.multiline-quotes = "single"
lint.flake8-type-checking.quote-annotations = true
lint.isort.force-single-line = true

[tool.pyproject-fmt]
column_width = 80
indent = 4

[tool.coverage.run]
branch = true
source = [ "src", "tests" ]

[tool.coverage.report]
show_missing = true
sort = "Cover"
skip_empty = true

[tool.mypy]
mypy_path = [ "src", "tests" ]
strict = true
warn_unused_configs = true
